requirements.txt
aiohttp==3.10.10
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pandas==2.2.3
google-generativeai==0.8.3
